contributors/gadams/incoming/cssom/stylesheet-interface-object.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD//XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>CSSOM - Interface Object - StyleSheet</title>
    <link rel="author" title="Glenn Adams" href="mailto:glenn@skynav.com"/>
    <link rel="help" href="http://www.w3.org/TR/cssom/#the-stylesheet-interface"/>
    <meta name="flags" content="dom"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
  </head>
  <body>
    <div id="log"></div>
    <div id="box"></div>
    <script><![CDATA[

      test(function(){

        assert_true(StyleSheet !== undefined);

      }, 'StyleSheet !== undefined');

      test(function(){

        assert_true(StyleSheet instanceof Object);

      }, 'StyleSheet instanceof Object');

      test(function(){

        assert_true(StyleSheet instanceof Function);

      }, 'StyleSheet instanceof Function');

    ]]>
    </script>
  </body>
</html>
